+++ internal/api/application_test.go
@@ -27,11 +27,18 @@ func TestApplicationAPI(t *testing.T) {
 		apiuser := NewUserAPI(lsCtx, validator)
 
 		Convey("When creating an application", func() {
-			createReq := &pb.CreateApplicationRequest{
-				Name:        "test-app",
-				Description: "A test application",
-			}
-			createResp, err := api.Create(ctx, createReq)
+			createResp, err := api.Create(ctx, &pb.CreateApplicationRequest{
+				Name:               "test-app",
+				Description:        "A test application",
+				IsABP:              true,
+				IsClassC:           true,
+				RxDelay:            1,
+				Rx1DROffset:        3,
+				RxWindow:           pb.RXWindow_RX2,
+				Rx2DR:              3,
+				AdrInterval:        20,
+				InstallationMargin: 5,
+			})
 			So(err, ShouldBeNil)
 			So(validator.ctx, ShouldResemble, ctx)
 			So(validator.validatorFuncs, ShouldHaveLength, 1)
@@ -45,9 +52,17 @@ func TestApplicationAPI(t *testing.T) {
 				So(validator.ctx, ShouldResemble, ctx)
 				So(validator.validatorFuncs, ShouldHaveLength, 1)
 				So(app, ShouldResemble, &pb.GetApplicationResponse{
-					Id:          createResp.Id,
-					Name:        "test-app",
-					Description: "A test application",
+					Id:                 createResp.Id,
+					Name:               "test-app",
+					Description:        "A test application",
+					IsABP:              true,
+					IsClassC:           true,
+					RxDelay:            1,
+					Rx1DROffset:        3,
+					RxWindow:           pb.RXWindow_RX2,
+					Rx2DR:              3,
+					AdrInterval:        20,
+					InstallationMargin: 5,
 				})
 			})
 
@@ -64,9 +79,17 @@ func TestApplicationAPI(t *testing.T) {
 				So(apps.Result, ShouldHaveLength, 1)
 				So(apps.TotalCount, ShouldEqual, 1)
 				So(apps.Result[0], ShouldResemble, &pb.GetApplicationResponse{
-					Id:          createResp.Id,
-					Name:        "test-app",
-					Description: "A test application",
+					Id:                 createResp.Id,
+					Name:               "test-app",
+					Description:        "A test application",
+					IsABP:              true,
+					IsClassC:           true,
+					RxDelay:            1,
+					Rx1DROffset:        3,
+					RxWindow:           pb.RXWindow_RX2,
+					Rx2DR:              3,
+					AdrInterval:        20,
+					InstallationMargin: 5,
 				})
 			})
 
@@ -112,9 +135,17 @@ func TestApplicationAPI(t *testing.T) {
 						So(apps.Result, ShouldHaveLength, 1)
 						So(apps.TotalCount, ShouldEqual, 1)
 						So(apps.Result[0], ShouldResemble, &pb.GetApplicationResponse{
-							Id:          createResp.Id,
-							Name:        "test-app",
-							Description: "A test application",
+							Id:                 createResp.Id,
+							Name:               "test-app",
+							Description:        "A test application",
+							IsABP:              true,
+							IsClassC:           true,
+							RxDelay:            1,
+							Rx1DROffset:        3,
+							RxWindow:           pb.RXWindow_RX2,
+							Rx2DR:              3,
+							AdrInterval:        20,
+							InstallationMargin: 5,
 						})
 					})
 
@@ -176,9 +207,17 @@ func TestApplicationAPI(t *testing.T) {
 
 			Convey("When updating the application", func() {
 				_, err := api.Update(ctx, &pb.UpdateApplicationRequest{
-					Id:          createResp.Id,
-					Name:        "test-app-updated",
-					Description: "An updated test description",
+					Id:                 createResp.Id,
+					Name:               "test-app-updated",
+					Description:        "An updated test description",
+					IsABP:              false,
+					IsClassC:           true,
+					RxDelay:            2,
+					Rx1DROffset:        4,
+					RxWindow:           pb.RXWindow_RX1,
+					Rx2DR:              1,
+					AdrInterval:        40,
+					InstallationMargin: 10,
 				})
 				So(err, ShouldBeNil)
 				So(validator.ctx, ShouldResemble, ctx)
@@ -190,9 +229,17 @@ func TestApplicationAPI(t *testing.T) {
 					})
 					So(err, ShouldBeNil)
 					So(app, ShouldResemble, &pb.GetApplicationResponse{
-						Id:          createResp.Id,
-						Name:        "test-app-updated",
-						Description: "An updated test description",
+						Id:                 createResp.Id,
+						Name:               "test-app-updated",
+						Description:        "An updated test description",
+						IsABP:              false,
+						IsClassC:           true,
+						RxDelay:            2,
+						Rx1DROffset:        4,
+						RxWindow:           pb.RXWindow_RX1,
+						Rx2DR:              1,
+						AdrInterval:        40,
+						InstallationMargin: 10,
 					})
 				})
 			})