+++ ui/src/views/nodes/ListNodes.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router';
 
+import Pagination from "../../components/Pagination";
 import NodeStore from "../../stores/NodeStore";
 import SessionStore from "../../stores/SessionStore";
 import ApplicationStore from "../../stores/ApplicationStore";
@@ -32,16 +33,17 @@ class ListNodes extends Component {
       application: {},
       nodes: [],
       isApplicationAdmin: false,
+      pageSize: 20,
+      pageNumber: 1,
+      pages: 1,
     };
 
     this.onDelete = this.onDelete.bind(this);
+    this.updatePage = this.updatePage.bind(this);
   }
 
   componentWillMount() {
-    NodeStore.getAll(this.props.params.applicationID, (nodes) => {
-      this.setState({nodes: nodes});
-    });
-
+    this.updatePage(this.props);
     ApplicationStore.getApplication(this.props.params.applicationID, (application) => {
       this.setState({application: application});
     });
@@ -54,6 +56,24 @@ class ListNodes extends Component {
       this.setState({
         isApplicationAdmin: (SessionStore.isAdmin() || SessionStore.isApplicationAdmin(this.props.params.applicationID)),
       });
+    });
+  }
+
+  componentWillReceiveProps(nextProps) {
+    this.updatePage(nextProps);
+  }
+
+
+  updatePage(props) {
+    const page = (props.location.query.page === undefined) ? 1 : props.location.query.page;
+  
+    NodeStore.getAll(this.props.params.applicationID, this.state.pageSize, (page-1) * this.state.pageSize, (totalCount, nodes) => {
+      this.setState({
+        nodes: nodes,
+        pageNumber: page,
+        pages: Math.ceil(totalCount / this.state.pageSize),
+      });
+      window.scrollTo(0, 0);
     });
   }
 
@@ -103,6 +123,7 @@ class ListNodes extends Component {
               </tbody>
             </table>
           </div>
+          <Pagination pages={this.state.pages} currentPage={this.state.pageNumber} pathname={`/applications/${this.props.params.applicationID}`} />
         </div>
       </div>
     );